# apps/admin-console/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SwissTIP local control API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assets": {
      "get": {
        "summary": "Assets",
        "operationId": "getAssets",
        "parameters": [
          {
            "name": "corpus_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corpus Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  },
                  "title": "Response Getassets"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload Asset",
        "operationId": "uploadAsset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assets/pilot": {
      "post": {
        "summary": "Load Pilot",
        "operationId": "loadPilotAssets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{identifier}/preview": {
      "get": {
        "summary": "Preview",
        "operationId": "previewAsset",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "getJobs",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "type": "array",
                  "title": "Response Getjobs"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Job",
        "operationId": "createJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{identifier}": {
      "get": {
        "summary": "Get Job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{identifier}/cancel": {
      "post": {
        "summary": "Cancel Job",
        "operationId": "cancelJob",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{identifier}/reviews": {
      "get": {
        "summary": "Reviews",
        "operationId": "getReviews",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Getreviews"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Review",
        "operationId": "addReview",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/releases/{identifier}/evidence": {
      "get": {
        "summary": "Evidence",
        "operationId": "getReleaseEvidence",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Getreleaseevidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "ActionRequest"
      },
      "Asset": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "corpus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Id"
          },
          "processing_eligible": {
            "type": "boolean",
            "title": "Processing Eligible",
            "default": true
          },
          "processing_reason": {
            "type": "string",
            "title": "Processing Reason",
            "default": ""
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "source_id",
          "filename",
          "sha256",
          "origin",
          "size",
          "created_at"
        ],
        "title": "Asset"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "request": {
            "additionalProperties": true,
            "type": "object",
            "title": "Request"
          },
          "log": {
            "type": "string",
            "title": "Log"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "result_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Sha256"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "request",
          "log",
          "result",
          "error",
          "cancel_requested",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "Job"
      },
      "JobRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "crawl",
              "plan",
              "extract"
            ],
            "title": "Kind"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Source Ids"
          },
          "asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Asset Ids"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "crawl_profile": {
            "type": "string",
            "const": "smoke",
            "title": "Crawl Profile",
            "default": "smoke"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "JobRequest"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
      "Preview": {
        "properties": {
          "extraction_profile": {
            "type": "string",
            "title": "Extraction Profile"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "characters": {
            "type": "integer",
            "title": "Characters"
          },
          "excluded_sections": {
            "type": "integer",
            "title": "Excluded Sections"
          }
        },
        "type": "object",
        "required": [
          "extraction_profile",
          "title",
          "language",
          "sections",
          "characters",
          "excluded_sections"
        ],
        "title": "Preview"
      },
...
      "ReviewRequest": {
        "properties": {
          "result_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Result Sha256"
          },
          "candidate_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Candidate Id"
          },
          "reviewer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reviewer"
          },
          "decision": {
            "type": "string",
            "enum": [
              "accept_draft",
              "reject",
              "needs_changes"
            ],
            "title": "Decision"
          },
          "notes": {
            "type": "string",
            "maxLength": 4000,
            "title": "Notes",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result_sha256",
          "candidate_id",
          "reviewer",
          "decision"
        ],
        "title": "ReviewRequest"
      },
...
      "UploadRequest": {
        "properties": {
          "source_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source Id"
          },
          "filename": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Filename"
          },
          "text": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "filename",
          "text"
        ],
        "title": "UploadRequest"
      },
...
    }
  }
}